Create shared types to reduce duplication between extension and pages
